rename training navigation hook
